Refactor to modularize more
Former-commit-id: 1d5ae90a514e575706d5a4ed333cb7bcea613fed
Former-commit-id: c66b2faff45e0b65779d7fed4d3355d859ae8e34
